requirements.txt
amqp==2.2.2
asgiref==3.2.3
billiard==3.5.0.5
celery==4.1.0
certifi==2019.11.28
Click==7.0
construct==2.5.3
diff-match-patch==20121119
Django==2.0.1
django-adminlte-admin==0.6.1
django-ckeditor==5.8.0
django-crispy-forms==1.8.0
django-dashing==0.3.4
django-formtools==2.1
django-haystack==2.8.1
django-import-export==0.6.1
django-js-asset==1.2.2
django-mdeditor==0.1.16
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==2.0.12
django-simple-captcha==0.5.6
djangorestframework==3.10.3
docopt==0.6.2
dukpy==0.2.0
et-xmlfile==1.0.1
Flask==1.1.1
Flask-Cors==3.0.8
future==0.16.0
gunicorn==19.10.0
httplib2==0.9.2
importlib-metadata==1.3.0
itsdangerous==1.1.0
jdcal==1.3
jieba==0.39
Jinja2==2.10.3
jupyter-echarts-pypkg==0.1.2
kombu==4.1.0
lml==0.0.2
macro==0.0.2
macropy3==1.1.0b2
Markdown==3.1.1
MarkupSafe==1.0
Micro==3.1.0
more-itertools==8.0.2
mysqlclient==1.4.6
numpy==1.18.0
odfpy==1.3.6
openpyxl==2.4.9
paho-mqtt==1.5.0
pandas==0.23.4
pandasql==0.7.3
pefile==2019.4.18
Pillow==5.0.0
prettytable==0.7.2
pyecharts==1.6.0
pyecharts-jupyter-installer==0.0.3
PyMySQL==0.9.3
pyperclip==1.5.26
python-dateutil==2.7.3
python-ptrace==0.9.4
pytz==2017.3
PyUserInput==0.1.9
PyYAML==3.12
records==0.5.3
setuptools-scm==3.3.3
simplejson==3.16.0
six==1.13.0
SQLAlchemy==1.3.11
sqlparse==0.3.0
tablib==0.12.1
unicodecsv==0.14.1
vine==1.3.0
Werkzeug==0.16.0
Whoosh==2.7.4
xlrd==1.1.0
xlwt==1.3.0
zipp==0.6.0
